tweaks types
